* 옵션을 로컬 파일로 저장하고 불러오도록 수정 * 롱노트가 싱크 변경 시 올바르지 않게 표기되는 버그 수정 * 채보 파일을 포맷에 맞추어 읽어올 수 있도록 수정 * 최고 기록을 저장하고, 불러올 수 있도록 수정 중

+++ Game/Manager/SoundManager.h
@@ -25,7 +25,7 @@ public:
 	static SoundManager*& Get();
 	static void Destroy();
 	void SetBGM(std::string const& title = "", UINT32 const& Highlight = 0);
-	void SetVolume(float const& volume);
+	void SetVolume(UINT const& volume);
 	void SetFadeIO();
 };
